+++ Libplanet/Blockchain/BlockChain.cs
@@ -717,6 +717,11 @@ namespace Libplanet.Blockchain
                     nonceDeltas[txSigner] = nonceDelta + 1;
                 }
 
+                if (evaluateActions)
+                {
+                    ExecuteActions(block, renderActions);
+                }
+
                 _rwlock.EnterWriteLock();
                 try
                 {
@@ -753,11 +758,6 @@ namespace Libplanet.Blockchain
             finally
             {
                 _rwlock.ExitUpgradeableReadLock();
-            }
-
-            if (evaluateActions)
-            {
-                ExecuteActions(block, renderActions);
             }
         }